+++ ExpandWorldSpawns/DropLoader.cs
@@ -47,6 +47,14 @@ public class Loader
   {
     data.biomes = DataManager.ToBiomes(data.biome, fileName);
     data.biomeAreas = DataManager.ToBiomeAreas(data.biomeArea, fileName);
+    if (data.log.Equals("none", StringComparison.OrdinalIgnoreCase))
+      data.logNone = true;
+    else if (data.log != "")
+      data.logObj = DataManager.ToPrefab(data.log, fileName);
+    if (data.stump.Equals("none", StringComparison.OrdinalIgnoreCase))
+      data.stumpNone = true;
+    else if (data.stump != "")
+      data.stumpObj = DataManager.ToPrefab(data.stump, fileName);
     foreach (var drop in data.drops)
     {
       drop.obj = DataManager.ToPrefab(drop.prefab, fileName);
@@ -57,7 +65,9 @@ public class Loader
     }
     return data;
   }
-  public static bool IsValid(Data data) => data.drops.All(d => d.obj != null);
+  public static bool IsValid(Data data) => data.drops.All(d => d.obj != null)
+    && (data.log == "" || data.logNone || data.logObj != null)
+    && (data.stump == "" || data.stumpNone || data.stumpObj != null);
 
   public static void Set(string yaml)
   {