+++ publish/CHANGELOG.md
@@ -1,5 +1,6 @@
 - v1.18
   - Adds biome filtering to drop definitions.
+  - Adds log and stump overrides for tree drops.
 
 - v1.17
   - Adds experimental drop data support.